% Generated by build script
\name{txgov_datasets}
\alias{txgov_datasets}
\title{List Texas Open Data datasets}
\usage{
txgov_datasets()
}
\arguments{
\item{limit}{Number of datasets (default 50)}
\item{query}{Optional search filter}
}
\value{tibble: id, name, category, description, rows_updated_at}
\description{List Texas Open Data datasets}
